requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
